fix GPU  memory copy bug

+++ src/core/sh-clipper.cpp
@@ -4,6 +4,7 @@
 #include "core/vertex.h"
 #include <algorithm>
 
+
 SHClipper::SHClipper()
 {
 }
@@ -64,7 +65,7 @@ void SHClipper::clip(vector<PointObject> &input,
                                     vector<PointObject> &output,
                                     Boundary b)
 {
-    for (int i = 0; i < input.size(); ++i) {
+    for (size_t i = 0; i < input.size(); ++i) {
         PointObject begin = input[i];
         PointObject end = input[(i+1)%input.size()];
